+++ grid2op/Exceptions/__init__.py
@@ -48,6 +48,11 @@ __all__ = [
     "ChronicsNotFoundError",
     "InsufficientData",
     "BackendError",
+    "DivergingPowerflow",
+    "IslandedGrid",
+    "IsolatedElement",
+    "DisconnectedLoad",
+    "DisconnectedGenerator",
     "PlotError",
     "OpponentError",
     "UsedRunnerError",
@@ -59,63 +64,70 @@ __all__ = [
 
 from grid2op.Exceptions.Grid2OpException import Grid2OpException
 
-from grid2op.Exceptions.EnvExceptions import EnvError
-from grid2op.Exceptions.EnvExceptions import IncorrectNumberOfLoads
-from grid2op.Exceptions.EnvExceptions import IncorrectNumberOfGenerators
-from grid2op.Exceptions.EnvExceptions import IncorrectNumberOfLines
-from grid2op.Exceptions.EnvExceptions import IncorrectNumberOfSubstation
-from grid2op.Exceptions.EnvExceptions import IncorrectNumberOfStorages
-from grid2op.Exceptions.EnvExceptions import IncorrectNumberOfElements
-from grid2op.Exceptions.EnvExceptions import IncorrectPositionOfLoads
-from grid2op.Exceptions.EnvExceptions import IncorrectPositionOfGenerators
-from grid2op.Exceptions.EnvExceptions import IncorrectPositionOfLines
-from grid2op.Exceptions.EnvExceptions import IncorrectPositionOfStorages
-from grid2op.Exceptions.EnvExceptions import UnknownEnv
-from grid2op.Exceptions.EnvExceptions import MultiEnvException
+from grid2op.Exceptions.EnvExceptions import (EnvError,
+                                              IncorrectNumberOfLoads,
+                                              IncorrectNumberOfGenerators,
+                                              IncorrectNumberOfLines,
+                                              IncorrectNumberOfSubstation,
+                                              IncorrectNumberOfStorages,
+                                              IncorrectNumberOfElements,
+                                              IncorrectPositionOfLoads,
+                                              IncorrectPositionOfGenerators,
+                                              IncorrectPositionOfLines,
+                                              IncorrectPositionOfStorages,
+                                              UnknownEnv,
+                                              MultiEnvException)
 
-from grid2op.Exceptions.IllegalActionExceptions import IllegalAction
-from grid2op.Exceptions.IllegalActionExceptions import OnProduction
-from grid2op.Exceptions.IllegalActionExceptions import VSetpointModified
-from grid2op.Exceptions.IllegalActionExceptions import ActiveSetPointAbovePmax
-from grid2op.Exceptions.IllegalActionExceptions import ActiveSetPointBelowPmin
-from grid2op.Exceptions.IllegalActionExceptions import OnLoad
-from grid2op.Exceptions.IllegalActionExceptions import OnLines
-from grid2op.Exceptions.IllegalActionExceptions import InvalidReconnection
-from grid2op.Exceptions.IllegalActionExceptions import (
-    UnitCommitorRedispachingNotAvailable,
-)
+from grid2op.Exceptions.IllegalActionExceptions import (IllegalAction,
+                                                        OnProduction,
+                                                        VSetpointModified,
+                                                        ActiveSetPointAbovePmax,
+                                                        ActiveSetPointBelowPmin,
+                                                        OnLoad,
+                                                        OnLines,
+                                                        InvalidReconnection,
+                                                        UnitCommitorRedispachingNotAvailable,
+                                                        )
 
-from grid2op.Exceptions.AmbiguousActionExceptions import NotEnoughGenerators
-from grid2op.Exceptions.AmbiguousActionExceptions import GeneratorTurnedOffTooSoon
-from grid2op.Exceptions.AmbiguousActionExceptions import GeneratorTurnedOnTooSoon
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidRedispatching
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidBusStatus
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidNumberOfObjectEnds
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidNumberOfLines
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidNumberOfGenerators
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidNumberOfLoads
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidCurtailment
-from grid2op.Exceptions.AmbiguousActionExceptions import UnrecognizedAction
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidLineStatus
-from grid2op.Exceptions.AmbiguousActionExceptions import InvalidStorage
-from grid2op.Exceptions.AmbiguousActionExceptions import AmbiguousAction
-from grid2op.Exceptions.AmbiguousActionExceptions import NonFiniteElement
-from grid2op.Exceptions.AmbiguousActionExceptions import AmbiguousActionRaiseAlert
+from grid2op.Exceptions.AmbiguousActionExceptions import (NotEnoughGenerators,
+                                                          GeneratorTurnedOffTooSoon,
+                                                          GeneratorTurnedOnTooSoon,
+                                                          InvalidRedispatching,
+                                                          InvalidBusStatus,
+                                                          InvalidNumberOfObjectEnds,
+                                                          InvalidNumberOfLines,
+                                                          InvalidNumberOfGenerators,
+                                                          InvalidNumberOfLoads,
+                                                          InvalidCurtailment,
+                                                          UnrecognizedAction,
+                                                          InvalidLineStatus,
+                                                          InvalidStorage,
+                                                          AmbiguousAction,
+                                                          NonFiniteElement,
+                                                          AmbiguousActionRaiseAlert)
 
 from grid2op.Exceptions.PowerflowExceptions import DivergingPowerFlow
 
-from grid2op.Exceptions.ObservationExceptions import BaseObservationError
-from grid2op.Exceptions.ObservationExceptions import NoForecastAvailable
-from grid2op.Exceptions.ObservationExceptions import SimulateError
-from grid2op.Exceptions.ObservationExceptions import SimulateUsedTooMuchThisStep
-from grid2op.Exceptions.ObservationExceptions import SimulateUsedTooMuchThisEpisode
+from grid2op.Exceptions.ObservationExceptions import (BaseObservationError,
+                                                      NoForecastAvailable,
+                                                      SimulateError,
+                                                      SimulateUsedTooMuchThisStep,
+                                                      SimulateUsedTooMuchThisEpisode)
 
-from grid2op.Exceptions.ChronicsExceptions import ChronicsError
-from grid2op.Exceptions.ChronicsExceptions import ChronicsNotFoundError
-from grid2op.Exceptions.ChronicsExceptions import InsufficientData
-from grid2op.Exceptions.handlers_exceptions import HandlerError
+from grid2op.Exceptions.ChronicsExceptions import (ChronicsError,
+                                                   ChronicsNotFoundError,
+                                                   InsufficientData,
+                                                   )
+from grid2op.Exceptions.handlers_exceptions import (HandlerError,
+                                                    )
 
-from grid2op.Exceptions.BackendExceptions import BackendError
+from grid2op.Exceptions.BackendExceptions import (BackendError,
+                                                  DivergingPowerflow,
+                                                  IslandedGrid,
+                                                  IsolatedElement,
+                                                  DisconnectedLoad,
+                                                  DisconnectedGenerator,
+                                                  )
 
 from grid2op.Exceptions.PlotExceptions import PlotError